+++ tests/ui-api-layer-acceptance-tests/installer/pod.go
@@ -4,7 +4,6 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/kyma-project/kyma/tests/ui-api-layer-acceptance-tests"
 	"github.com/kyma-project/kyma/tests/ui-api-layer-acceptance-tests/waiter"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -60,33 +59,4 @@ func (t *PodInstaller) WaitForPodRunning(k8sClient *corev1Type.CoreV1Client) err
 
 		return false, nil
 	}, podReadyTimeout)
-}
-
-func UPSBrokerPod(name string) *corev1.Pod {
-	return &corev1.Pod{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: name,
-			Labels: map[string]string{
-				"app": name,
-			},
-		},
-		Spec: corev1.PodSpec{
-			Containers: []corev1.Container{
-				{
-					Name:  name,
-					Image: tester.UPSBrokerImage,
-					Args: []string{
-						"--port",
-						"8080",
-						"-alsologtostderr",
-					},
-					Ports: []corev1.ContainerPort{
-						{
-							ContainerPort: 8080,
-						},
-					},
-				},
-			},
-		},
-	}
 }
