+++ tests/ui-api-layer-acceptance-tests/installer/service.go
@@ -6,7 +6,6 @@ import (
 	"github.com/kyma-project/kyma/tests/ui-api-layer-acceptance-tests/waiter"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/util/intstr"
 	corev1Type "k8s.io/client-go/kubernetes/typed/core/v1"
 )
 
@@ -56,27 +55,4 @@ func (t *ServiceInstaller) WaitForEndpoint(k8sClient *corev1Type.CoreV1Client) e
 
 		return true, nil
 	}, endpointReadyTimeout)
-}
-
-func UPSBrokerService(name string) *corev1.Service {
-	return &corev1.Service{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: name,
-			Labels: map[string]string{
-				"app": name,
-			},
-		},
-		Spec: corev1.ServiceSpec{
-			Selector: map[string]string{
-				"app": name,
-			},
-			Ports: []corev1.ServicePort{
-				{
-					Protocol:   corev1.ProtocolTCP,
-					Port:       80,
-					TargetPort: intstr.FromInt(8080),
-				},
-			},
-		},
-	}
 }
