+++ tests/ui-api-layer-acceptance-tests/upsbroker/broker.go
@@ -1,0 +1,19 @@
+package upsbroker
+
+import (
+	"github.com/kubernetes-incubator/service-catalog/pkg/apis/servicecatalog/v1beta1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+)
+
+func UPSServiceBroker(name, url string) *v1beta1.ServiceBroker {
+	return &v1beta1.ServiceBroker{
+		ObjectMeta: metav1.ObjectMeta{
+			Name: name,
+		},
+		Spec: v1beta1.ServiceBrokerSpec{
+			CommonServiceBrokerSpec: v1beta1.CommonServiceBrokerSpec{
+				URL: url,
+			},
+		},
+	}
+}
