+++ tests/ui-api-layer-acceptance-tests/upsbroker/pod.go
@@ -1,0 +1,36 @@
+package upsbroker
+
+import (
+	"github.com/kyma-project/kyma/tests/ui-api-layer-acceptance-tests"
+	corev1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+)
+
+func UPSBrokerPod(name string) *corev1.Pod {
+	return &corev1.Pod{
+		ObjectMeta: metav1.ObjectMeta{
+			Name: name,
+			Labels: map[string]string{
+				"app": name,
+			},
+		},
+		Spec: corev1.PodSpec{
+			Containers: []corev1.Container{
+				{
+					Name:  name,
+					Image: tester.UPSBrokerImage,
+					Args: []string{
+						"--port",
+						"8080",
+						"-alsologtostderr",
+					},
+					Ports: []corev1.ContainerPort{
+						{
+							ContainerPort: 8080,
+						},
+					},
+				},
+			},
+		},
+	}
+}
